Create OVFParser.py
This code is used to retrieve vector field data from an OVF2 file format. From an instance of the OVF class, a user can extract slices of the field along specified planes or cylindrical coordinates. The field is resampled using Scipy's Interpolation methods. This is designed to replicate some of the features of Ubermag's discretisedfield method.

OVF Info:
https://math.nist.gov/oommf/doc/userguide12b4/userguide/Vector_Field_File_Format_OV.html